monitor-data-pipeline/compute-node-agent/requirements.txt
# Compute Node Agent Requirements

# BPF/eBPF for system metrics collection
bcc>=0.18.0

# gRPC for streaming metrics to server
grpcio>=1.50.0
grpcio-tools>=1.50.0
protobuf>=4.21.0

# etcd client for configuration and heartbeat
etcd3>=0.12.0

# GPU monitoring (NVIDIA)
pynvml>=11.4.1

# Standard library enhancements
typing-extensions>=4.0.0
